removed custom css from hero
