--- web/components/Hero/Hero.css
@@ -1,22 +1,0 @@
-.title {
-    position: relative;
-    z-index: 1;
-    line-height: 5rem;
-    font-size: 4.8rem;
-}
-
-.title:hover::after {
-    width: 100%;
-}
-
-.title::after {
-    content: "";
-    background-color: #0ba750;
-    width: 25%;
-    height: 4px;
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    z-index: -1;
-    transition: width 250ms ease-in-out;
-}
